TextInput, TextArea: Change boolean hasError to validation status
These components can take in a status, which is one of four strings
representing different validation statuses. We currently only
add special styles for the error state, but we may add them for
other states in the future.

This patch removes the boolean `hasError` property and replces it
with a string property, `status`, that defaults to "default".

Change-Id: I30ba92fdfe3249186f1844ed87ebf30afdd1b08a