Table: Remove custom CSS-only icons
Previously, CSS-only icons had to be set up by the developer user
for the 4 pager buttons and 3 sort icons. These will now be set
automatically in the Table and TablePager CSS from Codex, so we
no longer have to set them per-table.

This patch removes the custom icons from Table and Pager and
adds the appropriate Codex CSS classes to the icon spans so
the icons can be set up in the Codex CSS. It also requires
v1.16.0 of `@wikimedia/codex` and friends in the sandbox to
ensure we have the new table icon CSS.

Bug: T379660
Change-Id: I69978ab1a1f31a109e7ab10e4cc216980feff413